+++ src/components/slotmachine/components/reel-content.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { Tile } from '../data';
+import { Tile } from '../../../store/data';
 
 interface ScProps {
   height: number;
